+++ learn-dx12/include/Rendering/Data/FrameGraph/GPUPresentGraphNode.hpp
@@ -18,7 +18,7 @@ public:
 
     void ImportRenderTarget(GPUFrameResource* renderTarget);
 
-    virtual void Process(UINT64 frameIndex) override;
+    virtual void Process(int frameIndex) override;
 
 protected:
     void TransitionRenderTargetState(int frameIndex);
